lambda/requirements.txt
Pillow>=9.5.0
numpy>=1.24.0
boto3>=1.28.0

